APPS CON ESTRUCTURA/AIFORENSICSCANNER/requirements.txt

gradio
torch
torchvision
scikit-learn
joblib
pillow
opencv-python
